Move styles into new file
